+++ src/main/java/com/example/GestionSoutenances/entities/Salle.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Entity
 @Data
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class Salle {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idSalle;
+    private int idSalle;
 
     private String nomSalle;
     private int capacite;
